test: calculate block hash test
