+++ src/block_hash.rs
@@ -3,3 +3,6 @@ pub mod event_commitment;
 pub mod receipt_commitment;
 pub mod state_diff_hash;
 pub mod transaction_commitment;
+
+#[cfg(test)]
+pub mod test_utils;
